moved contextMenuController into seperate file
